+++ src/utils.js
@@ -1,22 +1,20 @@
-import {adoptedSheetsRegistry, frame, OldCSSStyleSheet} from './shared';
+import {adoptedSheetsRegistry} from './shared';
+import {instanceOfStyleSheet} from './ConstructStyleSheet';
 
-export function instanceOfStyleSheet(instance) {
-  return (
-    instance instanceof OldCSSStyleSheet ||
-    instance instanceof frame.CSSStyleSheet
-  );
-}
+const importPattern = /@import/;
 
 export function checkAndPrepare(sheets, container) {
   const locationType = container === document ? 'Document' : 'ShadowRoot';
 
   if (!Array.isArray(sheets)) {
+    // document.adoptedStyleSheets = new CSSStyleSheet();
     throw new TypeError(
       `Failed to set the 'adoptedStyleSheets' property on ${locationType}: Iterator getter is not callable.`,
     );
   }
 
   if (!sheets.every(instanceOfStyleSheet)) {
+    // document.adoptedStyleSheets = [document.styleSheets[0]];
     throw new TypeError(
       `Failed to set the 'adoptedStyleSheets' property on ${locationType}: Failed to convert value to 'CSSStyleSheet'`,
     );
@@ -40,4 +38,18 @@ export function getAdoptedStyleSheet(location) {
       ? document
       : location,
   );
+}
+
+export function rejectImports(contents = '') {
+  const imports = contents.match(importPattern) || [];
+  let sheetContent = contents;
+  if (imports.length) {
+    console.warn(
+      '@import rules are not allowed here. See https://github.com/WICG/construct-stylesheets/issues/119#issuecomment-588352418'
+    );
+    imports.forEach(_import => {
+      sheetContent = sheetContent.replace(_import, '');
+    });
+  }
+  return sheetContent;
 }