+++ test/test.js
@@ -4,13 +4,16 @@ const paragraphStyles = new CSSStyleSheet();
 document.adoptedStyleSheets = [ headingStyles, paragraphStyles ];
 headingStyles.replace(` h1 {
     color: tomato;
-} `).then((sheet) => console.log(`${sheet} styles have been replaced.`));
+} `).then((sheet) => {
+  console.log(`${sheet} styles have been replaced.`)
+  console.assert(sheet === headingStyles, 'not the same', 'yass')
+});
 
 paragraphStyles.replaceSync(`p {
-  color: #1121212;
+  color: #ab2121;
   font-family: "Operator Mono", "Helvetica Neue";
 }`);
 
 setTimeout(() => {
-  headingStyles.addRule('*', 'font-family: Helvetica');
+  headingStyles.addRule('*', 'font-family: monospace');
 }, 1000);